refactor: relocate KBar components
