+++ next-docs/pages/_app.tsx
@@ -1,4 +1,4 @@
-import React, {ReactChild, ReactNode, useEffect, useMemo} from 'react';
+import React, {ReactChild, ReactNode} from 'react';
 import '../styles/globals.css';
 import '../styles/reset.css';
 import '../styles/themes.css';
@@ -7,15 +7,14 @@ import {
   KBarProvider,
   KBarPortal,
   KBarPositioner,
-  KBarAnimator,
-  useRegisterActions
+  KBarAnimator
 } from 'kbar';
 import { NextPage } from 'next';
 import Head from 'next/head';
+import { BarResults } from '../components/KBar/BarResults';
+import {BarSearch} from "../components/KBar/BarSearch";
 import Layout from '../components/Layout';
 import { DocsThemeProvider } from '../components/themes/DocsThemeProvider';
-import { BarResults } from './components/KBar/BarResults';
-import {BarSearch} from "./components/KBar/BarSearch";
 import type { AppProps } from 'next/app';
 
 type GetLayout = (page: ReactNode) => ReactChild & ReactNode;
